requirements.txt
requests==2.31.0
python-dotenv==1.0.0
configparser==6.0.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-watch==4.2.0
pytest-mock==3.12.0